refactor: Update timestamp parameters in PrometheusClient and tests to use int
- Changed the `start` and `end` parameters in the `query_range` method of PrometheusClient from float to int for improved precision.
- Updated related tests in `test_prometheus_client.py` and `test_etl_job.py` to reflect this change, ensuring consistency across the codebase.
- Adjusted mock responses and assertions to accommodate the new integer timestamp format.

+++ tests/test_etl_job.py
@@ -45,8 +45,8 @@ class DummyPromClient:
     def query_range(
         self,
         expr: str,
-        start: float,
-        end: float,
+        start: int,
+        end: int,
         step: str,
     ) -> dict[str, Any]:
         """Mock query_range method."""
